pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "watermark-cli"
version = "0.1.0"
description = "A CLI tool to add watermarks to images"
readme = "README.md"
authors = [{ name = "ShinChven", email = "shinchven@gmail.com"}]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "Pillow",
]
requires-python = ">=3.6"

[project.urls]
Homepage = "https://github.com/ShinChven/watermark-cli"

[project.scripts]
watermark = "watermark_cli.watermark:main"

[tool.setuptools]
packages = ["watermark_cli"]